backend/requirements.txt
asgiref==3.7.2
contourpy==1.1.0
cycler==0.11.0
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fonttools==4.40.0
kiwisolver==1.4.4
matplotlib==3.7.2
numpy==1.25.1
packaging==23.1
Pillow==10.0.0
PyJWT==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
